requirements.txt
Django==3.1.4
django-admin==2.0.1
djangorestframework==3.12.2
